Added gps_travel package
This gps_travel package is on the bot at ~/Desktop/Group4Proj/catkin_ws/src/gps_travel.
The group4.launch file is the preliminary launch file, it launches the master node, aria, gps and gps_travel. So far gps_travels only use is to start driving forward, but it is in the process of being worked on